I have written reports into Readme.md

+++ SQL/part1.sql
@@ -13,7 +13,7 @@ CREATE TABLE customers(
 CREATE TABLE phones(
     ID uuid PRIMARY KEY,
     CustomerID uuid,
-    Numbers INT [] NOT NULL,
+    Numbers BIGINT [] NOT NULL,
     Code VARCHAR(30) NOT NULL,
     FOREIGN KEY(CustomerID)
     REFERENCES customers(CustomerID)
@@ -25,7 +25,7 @@ CREATE TABLE addresses(
     Country VARCHAR(50) NOT NULL,
     City VARCHAR(50) NOT NULL,
     District VARCHAR(50) NOT NULL ,
-    PostalCode INT [] NOT NULL,
+    PostalCode BIGINT [] NOT NULL,
     FOREIGN KEY(CustomerID)
     REFERENCES customers(CustomerID)
     ON DELETE CASCADE
@@ -35,11 +35,11 @@ CREATE TABLE products(
     ID uuid PRIMARY KEY,
     CustomerID uuid,
     Name VARCHAR(50) NOT NULL,
-    Cost INT NOT NULL,
-    OrderNumber INT NOT NULL,
-    Amount INT NOT NULL,
+    Cost BIGINT NOT NULL,
+    OrderNumber BIGINT NOT NULL,
+    Amount BIGINT NOT NULL,
     Currency VARCHAR(50) NOT NULL,
-    Rating INT NOT NULL,
+    Rating BIGINT NOT NULL,
     FOREIGN KEY(CustomerID)
     REFERENCES customers(CustomerID)
     ON DELETE CASCADE
